requirements.txt
Flask==2.2.2
werkzeug==2.2.2
requests==2.27.1
pycryptodome==3.10.1
gunicorn==20.1.0
python-dotenv==1.0.1
python-telegram-bot==13.15
